+++ packages/integration-tester/src/validate.ts
@@ -68,7 +68,7 @@ export class TspRunner {
     this.dir = dir;
     this.suite = suite;
     this.#options = options;
-    this.#projectDirs = tspConfigDirs.slice(0, 10); // TODO: remove for test
+    this.#projectDirs = tspConfigDirs;
   }
 
   async run(): Promise<void> {
@@ -77,6 +77,7 @@ export class TspRunner {
       if (result.failureCount > 0) {
         throw new ValidationFailedError();
       }
+      return;
     }
     registerConsoleShortcuts(this);
     await this.rerunAll();