+++ packages/typespec-autorest/src/lib.ts
@@ -99,12 +99,6 @@ export interface AutorestEmitterOptions {
   "emit-lro-options"?: "none" | "final-state-only" | "all";
 
   /**
-   * Back-compat flag. If true, continue to emit `x-ms-client-flatten` in for some of the
-   * ARM resource properties.
-   * @deprecated
-   */
-  "arm-resource-flattening"?: boolean;
-  /**
    * Determines whether and how to emit schemas for common-types
    * @default "for-visibility-changes"
    */
@@ -228,13 +222,6 @@ const EmitterOptionsSchema: JSONSchemaType<AutorestEmitterOptions> = {
       default: "final-state-only",
       description:
         "Determine whether and how to emit x-ms-long-running-operation-options for lro resolution",
-    },
-    "arm-resource-flattening": {
-      type: "boolean",
-      nullable: true,
-      default: false,
-      description:
-        "Back-compat flag. If true, continue to emit `x-ms-client-flatten` in for some of the ARM resource properties.",
     },
     "emit-common-types-schema": {
       type: "string",